use String as discord id

+++ src/main/java/com/linkstart/api/repo/PlaylistRepo.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PlaylistRepo extends JpaRepository<Playlist, Long> {
+public interface PlaylistRepo extends JpaRepository<Playlist, Integer> {
     List<Playlist> findByMember(Member member);
     Page<Playlist> findByNameContaining(String name, Pageable pageable);
 }
